+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file.
 This project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- Changed functions operating on slices to return a `LagMatrix` struct
+  providing meta information about the generated matrix.
+
 ## [0.3.0] - 2023-09-02
 
 ### Added