+++ CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 This project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Added `Into<Box[T]>` for `TimeLag`.
+
+### Changed
+
+- Upgraded `ndarray` from 0.15 to 0.16.
+
 ## [0.4.0] - 2023-09-02
 
 ### Changed
@@ -42,6 +52,9 @@ This project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - 🎉 Initial release.
 
 [0.4.0]: https://github.com/sunsided/timelag-rs/releases/tag/0.4.0
+
 [0.3.0]: https://github.com/sunsided/timelag-rs/releases/tag/0.3.0
+
 [0.2.0]: https://github.com/sunsided/timelag-rs/releases/tag/0.2.0
+
 [0.1.0]: https://github.com/sunsided/timelag-rs/releases/tag/0.1.0