+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - Added `Into<Box[T]>` for `TimeLag`.
+- The crate is now explicitly labeled `forbid(unsafe)` unless the `unsafe` feature is enabled.
 
 ### Changed
 
@@ -16,6 +17,7 @@ This project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Internal
 
 - Add criterion benchmarks to the project.
+- Change `Into<_>` implementations to `From<_>` for broader support.
 
 ## [0.4.0] - 2023-09-02